fix: harden active paths per go-prod deep audit
Six-dimension pre-launch audit (integrations, rate limiting, data
leakage, UI wiring, perf, process). Active-path findings fixed;
dark-flag integration findings dispositioned to harden-before-flip
backlog. Tracker: .planning/reviews/go-prod-deep-audit-2026-07-10.md.

- leaks: payment-export idempotent replay re-selects safe shape;
  contractor.list omits ssnEncrypted; WHT certs gated payment:read;
  generic 5xx bodies (public-api + Fastify); KSeF credentialsRef and
  e-sign cross-org connectionId closed
- abuse: portal tRPC on strict limiter; magic-link per-email 5/15min
  fail-closed; per-subject caps on portal OCR + e-sign URL minting
- correctness: ClamAV scans full object (was first 4100 bytes);
  token-refresh lock/throttle repair; CSRF exemptions for QStash
  callback routes (/webhooks-outbound, /contract-health,
  /idp-deprovisioning)
- perf: 120s tx on import commit + leave finalize; economic-dependency
  scan O(A); peppol reconcile NOT EXISTS anti-join + ACTIVE precondition
- ops: job-meta for 3 uncovered crons; outbox FAILED-backlog gauge with
  region fan-out; CLAMAV/STORECOVE/CRON env in render.yaml + .env.example;
  API_URL fails fast on prod localhost default
- ui: translated error rendering + inline section error states
- tests: db mocks follow regional helpers; env prod API_URL cases
- wiki: 19 pages synced + log/hot + BM25

+++ packages/api/src/services/exports/index.ts
@@ -430,8 +430,8 @@ async function handleSpendByContractor(input: DispatchInput): Promise<DispatchRe
     SELECT
       c."legalName" AS "contractorName",
       COUNT(i.id)::int AS "invoiceCount",
-      COALESCE(SUM(i."amountToPayMinor")::int, 0) AS "totalMinor",
-      COALESCE(AVG(i."amountToPayMinor")::int, 0) AS "avgMinor",
+      COALESCE(SUM(i."amountToPayMinor")::bigint, 0)::bigint AS "totalMinor",
+      COALESCE(AVG(i."amountToPayMinor")::bigint, 0)::bigint AS "avgMinor",
       MAX(i."paidAt") AS "lastPaidAt"
     FROM "Invoice" i
     JOIN "Contractor" c ON c.id = i."contractorId"
@@ -487,7 +487,7 @@ async function handleSpendByTeam(input: DispatchInput): Promise<DispatchResult> 
       t.name AS "teamName",
       COUNT(DISTINCT c.id)::int AS "contractorCount",
       COUNT(i.id)::int AS "invoiceCount",
-      COALESCE(SUM(i."amountToPayMinor")::int, 0) AS "totalMinor"
+      COALESCE(SUM(i."amountToPayMinor")::bigint, 0)::bigint AS "totalMinor"
     FROM "Invoice" i
     JOIN "Contractor" c ON c.id = i."contractorId"
     LEFT JOIN "Team" t ON t.id = c."primaryTeamId"
